fix: remove dependency to @hello.nrfcloud.com/proto
This removes the dependency to @hello.nrfcloud.com/proto and
adds the source of the two utility files here directly.

The reason behind this is that the package pulls in other
dependencies, while in this repo we have the only dependency
(typebox) already.

The implementation of the code is also trivial so there is
a low chance on important upstream changes.