refactor: remove getHistoryTs()
This removes the getHistoryTs function and turns
it into smaller independent pieces.

The function depended unneccessary on a call to getSimDetailsFromCache()
which was ALWAYS called at the beginning of the function.
This is a clear indicator that it can be refactored in
only passing the result of the function call into the function (which removes the dependency).

In addition a utility function MaybeDate() is introduced
which allows to handle with potentially undefined Date or
date strings.

The data structure to store SIM usage details is converted
to an array so the timestamp information can be expressed
as a native Date object, instead of using a string.